add ttm support
