+++ packages/suite/l10n.config.json
@@ -1,0 +1,59 @@
+{
+    "extractedMessagesFilePattern": "translations/extractedMessages/**/*.json",
+    "localesOutputDir": "../suite-web/static/l10n/",
+    "outputDir": "translations/",
+    "project": {
+        "identifier": "trezor-wallet-test",
+        "filename": "master",
+        "apiKeyEnv": "CROWDIN_API_KEY"
+    },
+    "languages": [
+        "af",
+        "ar",
+        "bn",
+        "zh-CN",
+        "zh-TW",
+        "cs",
+        "nl",
+        "en",
+        "et",
+        "fr",
+        "de",
+        "el",
+        "he",
+        "hi",
+        "hu",
+        "id",
+        "it",
+        "ja",
+        "ko",
+        "no",
+        "fa",
+        "pl",
+        "pt-PT",
+        "ru",
+        "es-ES",
+        "sv-SE",
+        "tr",
+        "uk",
+        "vi"
+    ],
+    "langToFileNameMap": {
+        "en": "en",
+        "bn": "bn",
+        "cs": "cs",
+        "de": "de",
+        "el": "el",
+        "es-ES": "es",
+        "fr": "fr",
+        "id": "id",
+        "it": "it",
+        "ja": "ja",
+        "nl": "nl",
+        "pl": "pl",
+        "pt-PT": "pt",
+        "ru": "ru",
+        "uk": "uk",
+        "zh-CN": "zh"
+    }
+}
