+++ packages/suite/src/actions/languageActions.native.ts
@@ -1,0 +1,11 @@
+import { Dispatch } from '@suite/types';
+import * as SUITE from './constants/suite';
+
+// TODO
+export const fetchLocale = (locale: string) => (dispatch: Dispatch) => {
+    dispatch({
+        type: SUITE.SET_LANGUAGE,
+        locale,
+        messages: {},
+    });
+};
