+++ packages/translations-manager/package.json
@@ -1,7 +1,8 @@
 {
-  "name": "trezor-translations-manager",
+  "name": "@trezor/translations-manager",
   "version": "1.0.5",
   "main": "./lib/index.js",
+  "private": true,
   "repository": {
     "type": "git",
     "url": "https://github.com/trezor/trezor-translations-manager.git"
